Fill in placeholders in a SPARQL query based on provided parameters.

The `fill.query` function replaces placeholders in a SPARQL query with specified values based on taxonomy ID or scientific name. It utilizes the `get.sparql.values`, `get.named.graphs`, `query.by.name`, and `query.by.taxid` functions.

fill.query(query, taxid = NULL, sciName = NULL)

Arguments

query

A character string representing the SPARQL query with placeholders.

taxid

An optional parameter representing the Taxonomy ID to be used in the query.

sciName

An optional parameter representing the scientific name to be used in the query.

Value

A character string representing the filled-in SPARQL query with specified values.

Details

The function takes a SPARQL query with placeholders and optional parameters for Taxonomy ID and scientific name. It uses the `get.sparql.values` function to generate VALUES clauses for the provided parameters and replaces the placeholders in the query using `sprintf`. The function also includes information from `get.named.graphs`, `query.by.name`, and `query.by.taxid` in the resulting query.

Author

<Author 1>

Examples

# Example for filling in placeholders in a SPARQL query
sparql_query_template <- "SELECT * WHERE { %s }"
filled_query <- fill.query(sparql_query_template, taxid = c("GBIF:2130185", "NCBI:55786"))
#> Warning: 3 arguments not used by format 'SELECT * WHERE { %s }'
cat(filled_query)
#> SELECT * WHERE { 
#>     FROM NAMED <http://www.ontotext.com/explicit>
#>     FROM NAMED <http://www.ontotext.com/implicit>
#>     FROM NAMED <http://www.ontotext.com/skip-redundant-implicit>
#>    }