Retrieve interaction information based on Taxonomy ID or scientific name from a specific SPARQL endpoint.

The `get.interactions` function constructs and executes a SPARQL query to retrieve interaction information based on specified Taxonomy ID or scientific name. It utilizes the `fill.query`, `send.sparql`, and `format.result.df` functions.

get.interactions(taxid = NULL, sciName = NULL, endpoint = NULL)

Arguments

taxid

An optional parameter representing the Taxonomy ID(s) to be used in the query.

sciName

An optional parameter representing the scientific name(s) to be used in the query.

endpoint

A character string representing the SPARQL endpoint to send the query to.

Value

A data frame containing interaction information based on the specified parameters.

Details

The function constructs a SPARQL query template, fills in placeholders based on provided parameters using the `fill.query` function, sends the query to the specified SPARQL endpoint using `send.sparql`, and formats the result data frame using `format.result.df`. The resulting data frame contains the following information: query name, query ID, match name, match ID, interaction ID, interaction name, resource ID, resource name, reference, source, and inferred.

Author

<Author 1>

Examples

# Example for retrieving interaction information based on Taxonomy ID from a specific SPARQL endpoint
interaction_info <- get.interactions(taxid = c("GBIF:2130185", "NCBI:55786"), endpoint = "http://129.88.204.79:7200/repositories/gratin")
#> Error in req_perform(.): HTTP 404 Not Found.
print(interaction_info)
#> Error in eval(expr, envir, enclos): object 'interaction_info' not found