Send a SPARQL query to an endpoint and retrieve results.

The `send.sparql` function sends a SPARQL query to a specified endpoint and retrieves the results. The function also supports adding prefixes to the query using the `add_prefixes` function.

send.sparql(query, endpoint, verbose = FALSE)

Arguments

query

A character string representing the SPARQL query to be sent.

endpoint

A character string representing the SPARQL endpoint to send the query to.

verbose

Logical, indicating whether to print the SPARQL query before sending.

Value

A tibble containing the results of the SPARQL query.

Details

The function sends a GET request to the specified SPARQL endpoint with the provided query. It expects the response in the SPARQL Results JSON format and parses the results into a tibble. The function also checks the response type and raises an error if it's not in the expected format.

Author

<Author 1>

Examples

# Example for sending a SPARQL query to an endpoint
query <- "SELECT ?subject ?predicate ?object WHERE { ?subject ?predicate ?object } LIMIT 10"
endpoint <- "http://example.com/sparql"
result <- send.sparql(query, endpoint, verbose = TRUE)
#> PREFIX obo: <http://purl.obolibrary.org/obo/>
#> PREFIX dwc: <http://rs.tdwg.org/dwc/terms/>
#> PREFIX rdf: <http://www.w3.org/1999/02/22-rdf-syntax-ns#>
#> PREFIX rdfs: <http://www.w3.org/2000/01/rdf-schema#>
#> PREFIX xsd: <http://www.w3.org/2001/XMLSchema#>
#> PREFIX onto: <http://www.ontotext.com/>
#> PREFIX NCBITaxon: <http://purl.obolibrary.org/obo/NCBITaxon_>
#> PREFIX NCBI: <https://www.ncbi.nlm.nih.gov/Taxonomy/Browser/wwwtax.cgi?id=>
#> PREFIX RO: <http://purl.obolibrary.org/obo/RO_>
#> PREFIX SFWO: <http://purl.org/sfwo/SFWO_>
#> PREFIX EOL: <http://eol.org/pages/>
#> PREFIX ECOCORE: <http://purl.obolibrary.org/obo/ECOCORE_>
#> PREFIX OBI: <http://purl.obolibrary.org/obo/OBI_>
#> PREFIX BFO: <http://purl.obolibrary.org/obo/BFO_>
#> PREFIX GBIF: <http://www.gbif.org/species/>
#> PREFIX ITIS: <http://www.itis.gov/servlet/SingleRpt/SingleRpt?search_topic=TSN&search_value=>
#> PREFIX SILVA: <https://www.arb-silva.de/>
#> PREFIX IF: <http://www.indexfungorum.org/names/NamesRecord.asp?RecordID=>
#> PREFIX sesame: <http://www.openrdf.org/schema/sesame#>
#> PREFIX SLIME: <https://purl.slime.org/>
#> PREFIX dc: <http://purl.org/dc/elements/1.1/>
#> SELECT ?subject ?predicate ?object WHERE { ?subject ?predicate ?object } LIMIT 10
#> Error in req_perform(.): Failed to perform HTTP request.
#> Caused by error in `curl::curl_fetch_memory()`:
#> ! Failed to connect to example.com port 80 after 2 ms: Connection refused
print(result)
#> Error in eval(expr, envir, enclos): object 'result' not found